NXDRIVE-470: Don't delete unsynchronized pair state in case of move to an Office temp file

+++ nuxeo-drive-client/nxdrive/engine/watcher/local_watcher.py
@@ -729,7 +729,9 @@ class LocalWatcher(EngineWorker):
             if doc_pair is not None:
                 if doc_pair.pair_state == 'unsynchronized':
                     log.debug("Ignoring %s as marked unsynchronized", doc_pair.local_path)
-                    if evt.event_type == 'deleted' or evt.event_type == 'moved':
+                    if (evt.event_type == 'deleted'
+                        or evt.event_type == 'moved' and not is_office_temp_file(os.path.basename(evt.dest_path))):
+                        log.debug('Removing pair state for deleted or moved event: %r', doc_pair)
                         self._dao.remove_state(doc_pair)
                     return
                 self._handle_watchdog_event_on_known_pair(doc_pair, evt, rel_path)
